Arrow Key Input
I couldn't get input for Z to work. I could switch the input and use arrow keys to move Z, but any other key it would not read and I don't know why

Python would recognize that I was pressing the key but the command read in C# woudn't move. Strange but not an actual part of our simulation so I'll move on without it